refactor: PhaseSelect
